test: enable MainViewManager-integ-test
